agent_planner/requirements.txt
anthropic==0.49.0
crewai==0.108.0
python-dotenv==1.0.1
